Implementando impostos sobre transacao e ajustes nos testes automatizados

+++ src/main/java/br/com/alura/carteiraAPI/service/TransacaoService.java
@@ -1,5 +1,7 @@
 package br.com.alura.carteiraAPI.service;
 
+import java.math.BigDecimal;
+
 import javax.persistence.EntityNotFoundException;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
@@ -8,6 +10,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -31,23 +34,35 @@ public class TransacaoService {
 	private UsuarioRepository usuarioRepository;
 	
 	@Autowired
+	private CalculadoraDeImpostoService calculadoraDeImpostoService;
+	
+	@Autowired
 	private ModelMapper modelMapper;
 	
-	public Page<TransacaoDto> listar(Pageable paginacao) {
-		return transacaoRepository.findAll(paginacao).map(t -> modelMapper.map(t, TransacaoDto.class));
+	public Page<TransacaoDto> listar(Pageable paginacao, Usuario logado) {
+		return transacaoRepository.findAllByUsuario(logado, paginacao).map(t -> modelMapper.map(t, TransacaoDto.class));
 	}
 	
 	@Transactional
-	public TransacaoDto cadastrar(@RequestBody @Valid TransacaoFormDto dto) {
+	public TransacaoDto cadastrar(@RequestBody @Valid TransacaoFormDto dto, Usuario logado) {
 		Long idUsuario = dto.getUsuarioId();
 		
 		try {
 			Usuario usuario = usuarioRepository.getById(idUsuario);
+			
+			if (!logado.equals(usuario)) {
+				lancarErro403();
+			}
+			
 			Transacao transacao = modelMapper.map(dto, Transacao.class);
 			transacao.setId(null);
 			transacao.setUsuario(usuario);
 			
+			BigDecimal imposto = calculadoraDeImpostoService.calcular(transacao);
+			transacao.setImposto(imposto);
+			
 			transacaoRepository.save(transacao);
+			
 			return modelMapper.map(transacao, TransacaoDto.class);
 			
 		} catch (EntityNotFoundException e) {
@@ -56,19 +71,43 @@ public class TransacaoService {
 	}
 	
 	@Transactional
-	public TransacaoDto atualizar(@Valid AtualizacaoTransacaoFormDto dto) {
+	public TransacaoDto atualizar(@Valid AtualizacaoTransacaoFormDto dto, Usuario logado) {
 		Transacao transacao = transacaoRepository.getById(dto.getId());	
+		
+		if (!transacao.pertenceAo(logado)) {
+			lancarErro403();
+		}
+		
 		transacao.atualizarInformacoes(dto.getTicker(), dto.getPreco(), dto.getQuantidade(), dto.getData(), dto.getTipo());
 		return modelMapper.map(transacao, TransacaoDto.class);
 	}
 	
 	@Transactional
-	public void remover(Long id) {
-		transacaoRepository.deleteById(id);
+	public void remover(Long id, Usuario logado) {
+		
+		Transacao transacao = transacaoRepository
+				.findById(id)
+				.orElseThrow(() -> new EntityNotFoundException());
+		
+		if (transacao.pertenceAo(logado)) {
+			transacaoRepository.deleteById(id);
+		} else {
+			lancarErro403();
+		}
 	}
 
-	public TransacaoDetalhadaDto listarPorId(@NotNull Long id) {
+	public TransacaoDetalhadaDto listarPorId(@NotNull Long id, Usuario logado) {
+		
 		Transacao transacao = transacaoRepository.findById(id).orElseThrow(() -> new EntityNotFoundException());
+		
+		if (!transacao.pertenceAo(logado)) {
+			lancarErro403();
+		}
+		
 		return modelMapper.map(transacao, TransacaoDetalhadaDto.class);
+	}
+	
+	private void lancarErro403() {
+		throw new AccessDeniedException("Acesso negado!");
 	}
 }
